requirements.txt
pandas==1.2.4
seaborn==0.11.1
#pip
matplotlib==3.3.3
numpy==1.19.2
#jupyter
fastcsv
weka==1.0.6
sklearn
javabridge==1.0.19
python-weka-wrapper3==0.1.16
opyenxes==0.3.0
skfeature
arff==0.9
jsonpickle==1.4.2
pyyaml
dateparser
colorama==0.4.4
scipy
antlr4-python3-runtime==4.9.1
